requirements.txt
nltk==3.8.1
pandas==2.2.2
google-generativeai==0.8.3
matplotlib==3.8.4
python-dotenv==1.0.1
sacrebleu==2.4.0
pyarabic